refactor: separate router from app
