+++ src/layout/AppLayout.tsx
@@ -1,5 +1,4 @@
 import { Layout } from 'antd';
-// import { Outlet } from 'react-router-dom'
 import { HeaderPositionProvider } from '@/context/HeaderPositionProvider';
 import { Outlet } from 'react-router-dom';
