+++ src/layout/ProtectedLayout.tsx
@@ -1,7 +1,6 @@
 import { Layout } from 'antd';
 import AppHeader from './AppHeader';
 import { Content } from 'antd/es/layout/layout';
-// import { Outlet } from 'react-router-dom'
 import Sidebar from './Sidebar';
 import { HeaderPositionProvider } from '@/context/HeaderPositionProvider';
 import { Outlet } from 'react-router-dom';
