+++ tsconfig.json
@@ -28,5 +28,5 @@
     // "erasableSyntaxOnly": false,
     // "noUncheckedSideEffectImports": true
   },
-  "include": ["src", "vite.config.ts"]
+  "include": ["src", "vite.config.ts", "src/routes/Router.tsx"]
 }
